feat: new navigation for sidebar (#231)
Closes #213 
Closes #214 

- [x] Split grid view in two levels 

- [x] Ability to search globally or within a selected category

- [x] Added support to set/unset the ground from the sidebar

- [x] Removed grounds from edit project modal

- [x] Customize category thumbnail: 

The thumbnail of a category can be set by adding a `main: true` prop to an asset in that category. If more than one main asset is present, the first one will be used. If no main asset is present, default to use the first asset in the category for the thumbnail.

![new-sidebar](https://user-images.githubusercontent.com/2781777/53304321-96ccf900-3852-11e9-88ee-9c1c5cc37774.gif)