fix: Update models file (#3158)
* feat: Add utils to validate when a model size is valid

* fix: Validates the model size including the thumbnail size

* fix: Filesize comparation logic

* chore: Update @dcl/builder-client to version 6.0.1

* fix: Calculate final model size for wearables correctly

The code changes in this commit fix a bug in the calculation of the final model size for wearables. Previously, the old files were unnecessarily downloaded for emotes, which was causing incorrect size calculations. This commit updates the logic to only download the old files for wearables, excluding smart wearables and emotes.

* fix: Update max file size for thumbnails, wearables, and emotes